fix price in detail page

+++ Services/ProductService.cs
@@ -65,7 +65,7 @@ namespace CuaHangQuanAo.Services
                     PriceModifier = pv.PriceModifier,
                     StockQuantity = pv.StockQuantity,
                     AvailableInStorage = pv.Storages.Sum(s => s.Quantity ?? 0),
-                    FinalPrice = (int)(pv.Product.SellPrice + pv.Product.SellPrice * pv.PriceModifier)
+                    FinalPrice = (decimal)(pv.Product.SellPrice + pv.Product.SellPrice * pv.PriceModifier)
                 })
                 .Where(v => v.AvailableInStorage > 0) // Only show variants with stock
                 .OrderBy(v => v.Size)
@@ -117,7 +117,7 @@ namespace CuaHangQuanAo.Services
                     PriceModifier = pv.PriceModifier,
                     StockQuantity = pv.StockQuantity,
                     AvailableInStorage = pv.Storages.Sum(s => s.Quantity ?? 0),
-                    FinalPrice = (int)(pv.Product.SellPrice + pv.Product.SellPrice * pv.PriceModifier)
+                    FinalPrice = (decimal)(pv.Product.SellPrice + pv.Product.SellPrice * pv.PriceModifier)
                 })
                 .FirstOrDefaultAsync();
